fix merge conflict for friend&follower
