+++ moodBookProject/app/src/main/res/layout/user_item.xml
@@ -25,7 +25,6 @@
             android:layout_weight="1"
             android:padding="10dp"
             android:text="Username"
-            android:textColor="@android:color/black"
             android:textSize="25dp" />
 
     </LinearLayout>
